Update to 1.19.3, add event-based registry inspection and modification API (#4)
* Update to 1.19.3

* Do not infer several variable types at NMS RegistryHook

* Refactor registry hook logic to be event-based

The hooked codec now fires a custom RegistrySendEvent that is used
internally to implement biome overrides. In addition, this event can be
listened to by third-party plugins to do further advanced, low-level
registry manipulation and inspection.

* Set HOOK_ACTIVE flag as intended

This flag previously was never set, causing the /biomevisuals overrides
command to always report that the hooks were not active.

* Address var usage as per review comments

* Comment out the dimension hook type at HookType. Micro-optimize values map

* Replace event map entry null check in hooked codec by unmodifiable map

Because unmodified maps can't have entries added or removed (but they
can be modified), null checks are not necessary. This should be slightly
more performant in common cases, and alert users of wrong code promptly.

* Fix typo

* Get rid of some var. Rename HookType to RegistryType and expose in API

Exposing RegistryType to the public is necessary to make the event
payload more strongly typed.

* Refactor registry send events to use strongly typed objects

Several breaking changes to the API were made, and lots of code has to
be changed to account for deserializing and serializing registry data
from and to JSON.

This commit is a continuation of the previous commit.

* Address review comment

* Move events, add editorconfig

* Cleanups + general moving around

* Encapsulate registry data map usage on RegistrySendEvent

* Revert some biome effect data holder classes being refactored to records

* Make AmbientParticle constructor private

This provides better encapsulation, in line with how other constructors
for biome effect data holder classes are also private.

---------

Co-authored-by: Alejandro González <AlexTMjugador@users.noreply.github.com>
Co-authored-by: Owen <23108066+Owen1212055@users.noreply.github.com>