Associate the biomevisuals.command permission to /biomevisuals (#5)
When a command does not have any permission linked to it, it can be used
by any player on a server. However, it is undesirable to expose
technical commands like /biomevisuals to anyone in a public server.

To better support such scenario, let's restrict the usage of the
/biomevisuals command to those who have the `biomevisuals.command`
permission. In the absence of any permission plugins, such permission is
given to server operators.